requirements.txt
Flask==2.0.1
firebase-admin==5.0.3
python-dotenv==0.19.0
Flask-Session==0.5.0
Werkzeug==2.0.1 